+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * #4431: Fix monitors not showing the voiceline character.
 * #4431: Fix elevator videos not appearing.
 * Fix initial state for "variant" dropdown boxes on item properties pane.
+* Automatically patch `tools/toolsinvisible` to allow light to pass through.
 
 ### UCP-Relevant changes:
 * Add `SweepTrace` mode to `Collisions` result. This allows creating collisions